pyproject.toml
[project]
name = "media-manager"
version = "0.1.0"
description = "Automated music library maintenance for Navidrome/Plex/Symfonium"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "smbprotocol>=1.10.0",
    "mutagen>=1.47.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
media-manager = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
